Move `incrementName` into `str` and add test
